Fix bug where user password would not be saved
This was due to the fact that the UserProfile model did not have a password2
field. This caused the view to pass an empty string to the django hasher. Due
to the salting the resulting hash looked totally legit when you took a look
at the db. The user however could not log in.