Remove new_course view and related code
This only the first of many commits which will remove unnecessary data
management views (which are mostly untested).

We will be importing a list of courses during the setup of an INLOOP
instance. Moreover, the functionality to edit courses is already
provided by the Django admin interface out of the box.