Use `sudo` whitelist approach to run `docker`
Running docker is available to root or users in the group `docker`
only. To not run the complete web application with such elevated
privigeles, we use `sudo` to switch to group `docker` (the user
remains the same) only for this particular action.

In order for this to work, the administrator must whitelist the
command and the options used here in /etc/sudoers.

Whitelisting ensures several things:
  - only the user running gunicorn can call docker this way
  - it is not possible to call docker in another fashion, e.g.
    specifying other options or other Docker images

Documentation for the sudoers(5) entry will follow in a later commit.