Remove editor until we have a sane implementation
First of all the editor should be on a separate view and use a
separate template, because

 - nobody wants to scroll down the whole page to reach it
 - a realistic user would open another browser window showing the
   task description side-by-side

Second, the client-side implementation is extremely clumsy and lacks
certain important features such as saving without submitting as a
solution, adding a new file, etc.

Third, the view code for this blob is a real mess.