Use single quote strings where applicable
Conversion was automatically performed with a simple script based on
Python's tokenize module [1].

This coding convention is easy to follow, because when in doubt one can
just use `print(repr(s))` to see how Python would represent a string s
as a literal.  It prefers single quotes and uses double quotes for
strings that contain single quotes (e.g., "doesn't").

Docstrings must still be formatted with three double quotes.  This
is just a personal preference: the three double quotes create a clearly
visible "bar":

    """                         '''
    Some docstring.     vs.     Some docstring.
    """                         '''

Btw, it seems to me that single quotes are also better to reach on a
german keyboard when the ten finger system is applied strictly.

[1] https://docs.python.org/3/library/tokenize.html