Refactor with/except statements
Prefer words as variable names instead of single letters or acronyms,
for example `error` instead of `e` or `stream` instead of `fp`.

Furthermore, use the open(…) builtin for PathLike objects, because this
appears to be slightly more readable.

Normalized all `open(…)` calls to have a `mode` kwarg (but only when
it's not equal to 'r', which is the default) and got rid of the
`encoding` kwarg when it equals the default.

Reduce line length of `with` statements by importing ZipFile and
TemporaryDirectory directly.

Along the way, this fixed some bogus and unusal asserts for exceptions
in the tests.