Improve the scoped_id calculation
Unfortunately, it is not possible to embed the scoped_id calculation
as a subquery into the INSERT statement, because Django disallows F()
expressions in that case (see Django ticket 30129).

While trying to do that, I've come up with a better approach to
calculate the next scoped_id directly at the DB level using COALESCE.

Furthermore, I've extracted the calculation into its own method,
which will make it easier to inject IntegrityErrors in test cases
with mock.patch(…).