Load .env_* files explicitly in manage.py
Previously, pipenv and honcho did automatically pick up the
environment in the standard .env location, which, in our case, was a
symlink to .env.develop created during `make init`. This had several
drawbacks. For example, it wasn't possible to ad-hoc override vars
from the command line, and we needed a PIPENV_DOTENV_LOCATION
workaround for the test suite.