Add initial Content-Security-Policy (CSP) header
For style-src, font-src and img-src it was necessary to allow
'unsafe-inline' and loading from the data: scheme, respectively.
We still have a lot of inline styles lying around, which will not
change very soon. AnchorJS and the ACE.js editor use data: URIs
to load fonts and images.

Because X-Frame-Options is ignored when CSP is set, we include
an equivalently strict frame-ancestors rule.