Show alert when leaving with unsaved changes
Show a dialog box that warns the user that there are unsaved changes. The dialog only appears
when changes made to the code or new files were not saved yet. Used the "beforeunload" event of
the window element. Implementation contains some legacy methods to make it work in all common
browsers.

Fixes: #391