Fix a bug in the change detection initialization
If data.success or data.files aren't truthy, the HashComperator would
have been initialized with an empty list and a _mismatching_ empty
checksum, leading to a bogus immediate detection of a "change".

In practice, this made the save button active and prevented the user
from leaving the page without confirmation, given he/she never saved
any files for that task.