Add frontend logic and mock data for syntax check
The syntax check button now calls a function that will later do a request to check the syntax on server side. For now it just returns some generated mock data. The console above the editor can now be shown and hidden and display the result in a more readable way.