Made error handling for save/submit more robust
Added catch calls to fetch request, so that the user gets a message displayed when the fetch request itself failed, e.g. due to a missing internet connection. Furthermore added a param to saveFiles which indicates, whether it is a normal save action or a "save before submit" action. The latter now doesn't cause multiple alerts anymore.