Ignore PermissionError during removal of files
In a deployment setting, the file ownership doesn't permit the deletion
of the files associated with the models. Fixing this would require to
wrap the management command with `sudo`, which opens up a lot of other
problems on its own.

The easiest way is to invoke the following commmand regularly:

  ./manage.py unreferenced_files | grep media/solutions | sudo xargs rm

This will only run `xargs` and `rm` with root privileges.

Fixes: #366