Start runserver explicitly on localhost:8000 [skip ci]
Although this doesn't actually change which network interface the
server will listen on, it will make `runserver` print a server link
which is consistent with the django.contrib.sites fixture that we
ship, avoiding an annoying behavior on the admin pages.

The "view on site" links on admin detail pages always redirect to the
domain stored in the sites framework, regardless of the value of the
`Host` header that was used in the HTTP request. Following these
links seem to logout the user. However, that's not actually the case.
It's only that browsers do not regard 127.0.0.1 and localhost as the
same cookie domains and the session cookie is only valid on one of
them.