Require type annotations in non-test code
Add flake8-annotations linter plugin to check for the presence of type
annotations. This will intentionally break the build for the time being.

In the initial configuration for flake8-annotations, exclude the
complete tests/ directory and disable the (in my opinion) redundant
requirement of type annotations for `self` and `cls` method and
classmethod arguments (ANN101, ANN102).