Fix noxfile.py to use the proper format for pip constraints
We used to feed a full requirements.txt file to `pip install --constraint`
without issues because none of our packages used extras.

However, with the addition of pysocopg[binary] it appeared that we were
relying on the ondocumented and nowadays unsupported quirk in pip's
constraint files parser which allowed specifying package extras.